Refactor frontend file structure.
