+++ PhotoPlusFrontend/PhotoPlus/src/app/components/home/home.component.html
