+++ PhotoPlusFrontend/PhotoPlus/src/app/components/home/home.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { ProductService } from "../product.service"
-import { Product } from '../product';
+import { ProductService } from "../../services/product/product.service"
+import { Product } from '../../services/product/product';
 
 @Component({
   selector: 'app-home',
