+++ PhotoPlusFrontend/PhotoPlus/src/app/components/registration/registration.component.ts
