+++ PhotoPlusFrontend/PhotoPlus/src/app/models/link/link.spec.ts
