+++ PhotoPlusFrontend/PhotoPlus/src/app/models/link/link.ts
